Refactor authentication setup couse was a build error with next-auth
